Refactor the Ducks
