+++ arisia-remote/app/arisia/views/editDuck.scala.html
@@ -14,7 +14,7 @@
 @main(title) {
   <h1>@title</h1>
 
-  @b4.horizontal.form(arisia.controllers.routes.AdminController.duckModified(), "col-md-2", "col-md-10") { implicit hfc =>
+  @b4.horizontal.form(arisia.controllers.routes.DuckController.duckModified(), "col-md-2", "col-md-10") { implicit hfc =>
     @b4.text(duckForm("id"), Symbol("_label") -> "ID", Symbol("readonly") -> true)
     @b4.text(duckForm("imageUrl"), Symbol("_label") -> "Image URL")
     @b4.text(duckForm("altText"), Symbol("_label") -> "Alt Text")
